requirements.txt
Flask
python-dotenv
msal